Move seed data to own file. Prevent unnecessary for loop from running on each render.
