+++ src/App.js
@@ -1,7 +1,7 @@
 import { useState } from 'react';
-import namor from 'namor';
 import shortid from 'shortid';
 // own components
+import participantsData from './seedData';
 import Header from './Header';
 import AddParticipant from './Participant';
 import List from './List';
@@ -12,19 +12,6 @@ import '../node_modules/@fortawesome/fontawesome-free/css/all.css';
 function App() {
 
     const [sortedBy, setSortedBy] = useState({ column : 'name', ascending: true });
-
-    // generate data
-    const newPerson = (n) => {
-        const name = namor.generate({ words: 1, saltLength: 4 });
-        const phone = '0'+ Math.floor(Math.random() * 1000000000);
-        return { id: n, name, email: name + '@email.com', phone };
-    }
-
-    let participantsData = [];
-    for (let i = 0; i < 20; i++) {
-        participantsData.push(newPerson(shortid.generate()));
-    }
-
     const [persons, setPersons] = useState(participantsData);
 
     const addPerson = (personObject) => {
